--- src/hooks/useCartManagement.ts
@@ -1,124 +1,0 @@
-import { useEffect, useState } from 'react';
-import postCartItem from '../api/postCartItem';
-import deleteCartItem from '../api/deleteCartItem';
-import patchCartItem from '../api/patchCartItem';
-import { AddCartItemType, cartDataType } from '../types/cartItem';
-import { CART_LIMIT } from '../constants/carts';
-import { useToast } from './useToast';
-
-function useCartManagement({
-  refetchCarts,
-  carts,
-}: {
-  refetchCarts: () => Promise<void>;
-  carts: cartDataType[] | null;
-}) {
-  const [isErrorAddCartItem, setIsErrorAddCartItem] = useState(false);
-  const [isErrorDeleteCartItem, setIsErrorDeleteCartItem] = useState(false);
-  const [isErrorUpdateCartItem, setIsErrorUpdateCartItem] = useState(false);
-  const [errorAddCartItemMessage, setErrorAddCartItemMessage] = useState<string | null>(null);
-  const [errorDeleteCartItemMessage, setErrorDeleteCartItemMessage] = useState<string | null>(null);
-  const [errorUpdateCartItemMessage, setErrorUpdateCartItemMessage] = useState<string | null>(null);
-  const [isOverItemCounts, setIsOverItemCounts] = useState(false);
-  const [itemCount, setItemCount] = useState(0);
-
-  useToast(errorAddCartItemMessage, 'error');
-  useToast(errorDeleteCartItemMessage, 'error');
-  useToast(errorUpdateCartItemMessage, 'error');
-  useToast(
-    isOverItemCounts ? `장바구니는 최대 ${CART_LIMIT}개의 상품을 담을 수 있습니다.` : null,
-    'error',
-  );
-
-  const handleAddCartItem = async ({ productId, quantity }: AddCartItemType) => {
-    if (itemCount >= CART_LIMIT) {
-      setIsOverItemCounts(true);
-      return;
-    }
-
-    try {
-      await postCartItem({
-        productId,
-        quantity,
-      });
-      await refetchCarts();
-      setErrorAddCartItemMessage(null);
-      setIsErrorAddCartItem(false);
-    } catch (error) {
-      setIsErrorAddCartItem(true);
-      setErrorAddCartItemMessage(
-        error instanceof Error
-          ? error.message
-          : '장바구니에 상품을 추가하는 중 오류가 발생했습니다',
-      );
-    }
-  };
-
-  const handleUpdateCartItem = async ({
-    productId,
-    quantity,
-  }: {
-    productId: number;
-    quantity: number;
-  }) => {
-    const cartId = carts?.filter((cart) => cart.product.id === productId)[0].id || 0;
-
-    try {
-      await patchCartItem({ cartId, quantity });
-      await refetchCarts();
-      setErrorUpdateCartItemMessage(null);
-      setIsErrorUpdateCartItem(false);
-    } catch (error) {
-      setIsErrorUpdateCartItem(true);
-      setErrorUpdateCartItemMessage(
-        error instanceof Error
-          ? error.message
-          : '장바구니 수량을 업데이트하는 중 오류가 발생했습니다',
-      );
-    }
-  };
-
-  const handleDeleteCartItem = async ({ productId }: { productId: number }) => {
-    const cartId = carts?.filter((cart) => cart.product.id === productId)[0].id || 0;
-
-    try {
-      await deleteCartItem({ cartId });
-      await refetchCarts();
-      setErrorDeleteCartItemMessage(null);
-      setIsErrorDeleteCartItem(false);
-    } catch (error) {
-      setIsErrorDeleteCartItem(true);
-      setErrorDeleteCartItemMessage(
-        error instanceof Error
-          ? error.message
-          : '장바구니에 상품을 삭제하는 중 오류가 발생했습니다',
-      );
-    }
-  };
-
-  useEffect(() => {
-    if (carts) {
-      setItemCount(new Set(carts?.map((cart) => cart.product.id)).size);
-    }
-  }, [carts]);
-
-  useEffect(() => {
-    if (!isErrorAddCartItem) setIsOverItemCounts(false);
-  }, [isErrorAddCartItem]);
-
-  return {
-    handleAddCartItem,
-    handleUpdateCartItem,
-    handleDeleteCartItem,
-    isErrorAddCartItem,
-    isErrorDeleteCartItem,
-    isErrorUpdateCartItem,
-    isOverItemCounts,
-    itemCount,
-    errorAddCartItemMessage: errorAddCartItemMessage || '',
-    errorDeleteCartItemMessage: errorDeleteCartItemMessage || '',
-    errorUpdateCartItemMessage: errorUpdateCartItemMessage || '',
-  };
-}
-
-export default useCartManagement;
